Refactor AddTasksAppealsView
Move tasks related to an appeal into its own component.
