+++ client/app/queue/correspondence/intake/components/TasksAppeals/AddTasksAppealsView.jsx
@@ -1,16 +1,9 @@
 import React, { useEffect, useState } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import Checkbox from '../../../../../components/Checkbox';
-import Button from '../../../../../components/Button';
-import TaskNotRelatedToAppeal from '../TaskNotRelatedToAppeal';
 import RadioField from '../../../../../components/RadioField';
-import CaseListTable from '../../../../CaseListTable';
-import ApiUtil from '../../../../../util/ApiUtil';
-import { prepareAppealForStore } from '../../../../utils';
-import LoadingContainer from '../../../../../components/LoadingContainer';
-import { LOGO_COLORS } from '../../../../../constants/AppConstants';
-import { setRelatedTaskAppeals } from '../../../correspondenceReducer/correspondenceActions';
+import AddAppealRelatedTaskView from './AddAppealRelatedTaskView';
 
 const mailTasksLeft = [
   'Change of address',
@@ -32,13 +25,9 @@ const existingAppealAnswer = [
 ];
 
 export const AddTasksAppealsView = (props) => {
-  const [appeals, setAppeals] = useState([]);
-  const taskRelatedAppeals = useSelector(state => state.intakeCorrespondence.relatedTaskAppeals);
-  const [loading, setLoading] = useState(false);
+  const taskRelatedAppeals = useSelector((state) => state.intakeCorrespondence.relatedTaskAppeals);
   const [relatedToExistingAppeal, setRelatedToExistingAppeal] = useState(false);
   const [existingAppealRadio, setExistingAppealRadio] = useState('2');
-
-  const dispatch = useDispatch();
 
   const selectYes = () => {
     if (existingAppealRadio === '2') {
@@ -53,111 +42,6 @@ export const AddTasksAppealsView = (props) => {
       setRelatedToExistingAppeal(false);
     }
   };
-
-  const checkboxOnChange = useCallback((id, isChecked) => {
-    if (isChecked) {
-      dispatch(setRelatedTaskAppeals([...taskRelatedAppeals, id]));
-    } else {
-      const selected = taskRelatedAppeals.filter((checkboxId) => checkboxId !== id);
-
-      dispatch(setRelatedTaskAppeals(selected));
-    }
-  }, [taskRelatedAppeals]);
-
-  const clickAddTask = () => {
-    props.setAddTasksVisible(true);
-    const currentTask = [...props.unrelatedTasks];
-    const randNum = Math.floor(Math.random() * 1000000);
-
-    currentTask.push({ Object: randNum, Task: '', Text: '', SelectedTaskType: -1, SelectedTaskName: '' });
-    props.setUnrelatedTasks(currentTask);
-    props.disableContinue(false);
-  };
-
-  const removeTaskAtIndex = (index) => {
-    const currentTask = [...props.unrelatedTasks];
-    const newTask = currentTask.filter((item, i) => index !== i);
-
-    props.setUnrelatedTasks(newTask);
-    if (currentTask.length >= 2) {
-      props.disableContinue(true);
-    }
-    if (newTask.length === 0 || newTask.isNull) {
-      props.setAddTasksVisible(false);
-      props.disableContinue(true);
-    }
-  };
-
-  const [, setInstructionText] = useState('');
-
-  const checkContinueStatus = () => {
-    const currentTask = [...props.unrelatedTasks];
-
-    let continueEnabled = true;
-
-    currentTask.forEach((selectedTask) => {
-      if (selectedTask.SelectedTaskType === -1 || selectedTask.Text === '') {
-        // the condition is met
-        continueEnabled = false;
-      }
-    });
-
-    props.disableContinue(continueEnabled);
-  };
-
-  const handleChangeTaskTypeandText = (selectedOption, newText, index) => {
-    const newType = selectedOption.value;
-    const newLabel = selectedOption.label;
-    const currentTask = [...props.unrelatedTasks];
-
-    currentTask[index].SelectedTaskType = newType;
-    currentTask[index].SelectedTaskName = newLabel;
-    props.setUnrelatedTasks(currentTask);
-    setInstructionText(newText);
-    currentTask[index].Task = newType;
-    currentTask[index].Text = newText.trimStart();
-  };
-
-  useEffect(() => {
-    checkContinueStatus();
-  });
-
-  useEffect(() => {
-    // Only fetch if user indicates appeals data is needed
-    if (relatedToExistingAppeal === false) {
-      return;
-    }
-
-    // Don't refetch (use cache)
-    if (appeals.length) {
-      return;
-    }
-
-    // Visually indicate that we are fetching data
-    setLoading(true);
-
-    ApiUtil.get(`/queue/correspondence/${props.correspondenceUuid}/veteran`).
-      then((vetResponse) => {
-        const veteranFileNumber = vetResponse.body.file_number;
-
-        ApiUtil.get('/appeals', { headers: { 'case-search': veteranFileNumber } }).
-          then((appealResponse) => {
-            const appealsForStore = prepareAppealForStore(appealResponse.body.appeals);
-
-            const appealArr = [];
-
-            for (const appealGuid in appealsForStore.appeals) {
-              if (Object.prototype.hasOwnProperty.call(appealsForStore.appeals, appealGuid)) {
-                appealArr.push(appealsForStore.appeals[appealGuid]);
-              }
-            }
-
-            setAppeals(appealArr);
-            setLoading(false);
-          });
-      }
-      );
-  }, [relatedToExistingAppeal]);
 
   useEffect(() => {
     // If user has selected appeals, enable continue
@@ -205,60 +89,10 @@ export const AddTasksAppealsView = (props) => {
         <p style={{ marginTop: '0rem', marginBottom: '2rem' }}>
           Add new tasks related to this correspondence or to an appeal not yet created in Caseflow.
         </p>
-        <div />
+        <div>
+          <p>Placeholder</p>
+        </div>
 
-        {/* This button will activate the New Tasks section */}
-        {!props.addTasksVisible && <Button
-          type="button"
-          onClick={clickAddTask}
-          disabled={props.unrelatedTasks.length === 2}
-          name="addTaskOpen"
-          classNames={['cf-left-side']}>
-            + Add tasks
-        </Button>}
-
-        {/* This is the New Tasks section. Tasks will show next to each other in line. */}
-        {props.addTasksVisible &&
-        <div className="gray-border"
-          style={{ padding: '0rem 0rem', display: 'flex', flexWrap: 'wrap', flexDirection: 'column' }}>
-          <div style={{ width: '100%', height: 'auto', backgroundColor: 'white', paddingBottom: '3rem' }}>
-            <div style={{ backgroundColor: '#f1f1f1', width: '100%', height: '50px', paddingTop: '1.5rem' }}>
-              <b style={{
-                verticalAlign: 'center',
-                paddingLeft: '2.5rem',
-                paddingTop: '1.5rem',
-                border: '0',
-                paddingBottom: '1.5rem',
-                paddingRigfht: '5.5rem'
-              }}>New Tasks</b>
-            </div>
-            <div style={{ width: '100%', height: '3rem' }} />
-            <div style={{ display: 'flex', flexWrap: 'wrap' }}>
-              { props.unrelatedTasks && props.unrelatedTasks.map((currentTask, i) => (
-                <TaskNotRelatedToAppeal
-                  key={currentTask.Object}
-                  removeTask={() => removeTaskAtIndex(i)}
-                  handleChangeTaskType={(selectedOption, newText) =>
-                    handleChangeTaskTypeandText(selectedOption, newText, i)}
-                  taskType={currentTask.SelectedTaskType}
-                  taskText={currentTask.Text}
-                />
-              ))}
-
-            </div>
-            <div style={{ padding: '2.5rem 2.5rem' }} >
-              <Button
-                type="button"
-                onClick={clickAddTask}
-                disabled={props.unrelatedTasks.length === 2}
-                name="addTasks"
-                classNames={['cf-left-side']}>
-                  + Add tasks
-              </Button>
-            </div>
-          </div>
-        </div>}
-        <br></br>
         <h2>Tasks related to an existing Appeal</h2>
         <p>Is this correspondence related to an existing appeal?</p>
         <RadioField
@@ -267,20 +101,10 @@ export const AddTasksAppealsView = (props) => {
           options={existingAppealAnswer}
           onChange={existingAppealRadio === '2' ? selectYes : selectNo}
         />
-        {existingAppealRadio === '1' && loading &&
-          <LoadingContainer color={LOGO_COLORS.QUEUE.ACCENT}>
-            <div className="loading-div">
-            </div>
-          </LoadingContainer>
-        }
-        {existingAppealRadio === '1' && !loading &&
-          <div className="gray-border" style={{ marginBottom: '2rem', padding: '3rem 4rem' }}>
-            <CaseListTable
-              appeals={appeals}
-              showCheckboxes
-              checkboxOnChange={checkboxOnChange}
-            />
-          </div>
+        {existingAppealRadio === '1' &&
+          <AddAppealRelatedTaskView
+            correspondenceUuid={props.correspondenceUuid}
+          />
         }
       </div>
     </div>
@@ -288,11 +112,6 @@ export const AddTasksAppealsView = (props) => {
 };
 
 AddTasksAppealsView.propTypes = {
-  addTasksVisible: PropTypes.bool,
-  setAddTasksVisible: PropTypes.func,
-  disableContinue: PropTypes.func,
-  unrelatedTasks: PropTypes.arrayOf(Object),
-  setUnrelatedTasks: PropTypes.func,
   correspondenceUuid: PropTypes.string.isRequired,
   onContinueStatusChange: PropTypes.func.isRequired
 };
